refactor: comprehensive codebase cleanup — unified backends, lint enforcement, and simplification

+++ src/args.rs
@@ -1,14 +1,14 @@
 use crate::data_access::DataAccessMode;
 use clap::{Parser, ValueEnum};
 
-#[derive(ValueEnum, Debug, Clone, Copy, PartialEq, Default)]
+#[derive(ValueEnum, Debug, Clone, Copy, PartialEq, Eq, Default)]
 pub enum Interface {
     #[default]
     Ollama,
     Llamacpp,
 }
 
-#[derive(ValueEnum, Debug, Clone, Copy, PartialEq, Default)]
+#[derive(ValueEnum, Debug, Clone, Copy, PartialEq, Eq, Default)]
 pub enum OverwritePolicy {
     /// Skip any asset that already has a description
     #[default]
@@ -21,6 +21,7 @@ pub enum OverwritePolicy {
 
 #[derive(Parser, Debug, Clone)]
 #[command(author, version, about, long_about = None)]
+#[expect(clippy::struct_excessive_bools)]
 pub struct Args {
     /// Enable folder monitoring mode
     #[arg(short, long)]
@@ -44,13 +45,13 @@ pub struct Args {
     /// Path to Immich root directory (containing upload/, thumbs/ folders)
     #[arg(long, default_value = "/var/lib/immich")]
     pub immich_root: String,
-    /// PostgreSQL connection string (used only in database mode)
+    /// `PostgreSQL` connection string (used only in database mode)
     #[arg(
         long,
         default_value = "host=localhost user=postgres dbname=immich password=your_password"
     )]
     pub postgres_url: String,
-    /// Data access mode: database (direct PostgreSQL) or api (Immich REST API)
+    /// Data access mode: database (direct `PostgreSQL`) or api (Immich REST API)
     #[arg(short, long, value_enum, default_value = "database")]
     pub data_access_mode: DataAccessMode,
     /// Immich API base URL (required when using api access mode)
@@ -67,7 +68,7 @@ pub struct Args {
     pub immich_api_keys: Vec<String>,
     /// API poll interval in seconds (for Immich API mode)
     #[arg(long, default_value_t = 10)]
-    pub api_poll_interval: u64,
+    pub api_poll_interval: u32,
     /// Ollama model name for image analysis
     #[arg(long, default_value = "qwen3-vl:4b-thinking-q4_K_M")]
     pub model_name: String,
